+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Added
 - Command line tool `bnid`.
+- `min/maxEncodedIdBytes` utiltiies. Useful for variable length encodings such
+  as `base58`.
 
 ## 1.0.0 - 2020-05-07